backend/backend/urls.py
from django.contrib import admin
from django.urls import path, include
from api.views import CreateUserView, user_list, post_list
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
from django.conf import settings
from django.conf.urls.static import static

urlpatterns = [
    path("admin/", admin.site.urls),
    path("api/user/register/", CreateUserView.as_view(), name="register"),
    path('api/users/', CreateUserView.as_view(), name='user-create'),
    path('api/users/list/', user_list, name='user-list'),
    path("api/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
    path("api-auth/", include("rest_framework.urls")),
    path("api/", include("api.urls")), # include the api app urls
    path("api/posts/list", post_list, name="post-list")
]


if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)